backend/requirements.txt
fastapi
uvicorn
google-generativeai
pinecone
python-multipart
pypdf
python-dotenv
